Attempted delete on SPA but
failed.

+++ bookmarks/views.py
@@ -1,14 +1,10 @@
-from django.shortcuts import render
+from django.shortcuts import render, redirect, get_object_or_404
 from .forms import BookmarkForm
 from .models import Bookmark, PersonalBookmark
-from django.shortcuts import redirect, get_object_or_404
+from django.http import HttpResponseRedirect
 
 
-def index(request, bookmark_id=None):
-    if bookmark_id:
-        print('bookmark id exists')
-        delete_bookmark(request, bookmark_id)
-
+def index(request):
     if request.method == 'POST':
         form = BookmarkForm(request.POST)
         if form.is_valid():
@@ -32,31 +28,42 @@ def index(request, bookmark_id=None):
     return render(request, 'bookmarks/index.html', context)
 
 
-def delete_bookmark(request, bookmark_id):
-    print('bookmark_id', bookmark_id)
+def delete(request, bookmark=-1):
+    print('delete begin')
+    print('bookmark', bookmark)
+
+    if bookmark == -1:
+        return index(request)
+
+    context = {}
+
     bookmark = get_object_or_404(Bookmark, id=bookmark_id)
 
     if (request.method == 'POST'):
-        form = DeleteNewForm(request.POST, instance=bookmark)
+        form = BookmarkForm(request.POST, instance=bookmark)
 
         if form.is_valid():
             bookmark.delete()
-            return HttpResponseRedirect("/bookmarks")
-
+            return HttpResponseRedirect(reverse('bookmarks:index', args=(bookmark)))
     else:
-        form = DeleteNewForm(instance=bookmark)
+        form = BookmarkForm(instance=bookmark)
 
-    # context = {}
+    context = {}
 
-    # context['bookmarks'] = Bookmark.objects.exclude(
-    #     id__in=PersonalBookmark.objects.values_list('id'))
+    # context['invalid_form'] = form
 
-    # if request.user.is_anonymous:
-    #     context['personal_bookmarks'] = PersonalBookmark.objects.none()
-    # else:
-    #     context['personal_bookmarks'] = PersonalBookmark.objects.filter(
-    #         user=request.user)
+    context['bookmarks'] = Bookmark.objects.exclude(
+        id__in=PersonalBookmark.objects.values_list('id'))
 
-    # context['form'] = BookmarkForm()
+    if request.user.is_anonymous:
+        context['personal_bookmarks'] = PersonalBookmark.objects.none()
+    else:
+        context['personal_bookmarks'] = PersonalBookmark.objects.filter(
+            user=request.user)
 
-    # return render(request, 'bookmarks/index.html', context)
+    context['form'] = BookmarkForm()
+
+    return HttpResponseRedirect('/bookmarks')
+
+    print('delete end')
+    return render(request, 'bookmarks/index.html', context)
